+++ routes/api/contacts-router.js
@@ -1,93 +1,17 @@
 import express from 'express';
-import Joi from 'joi';
 
-import contactsServise from '../../models/contacts.js';
-import { HttpError } from '../../helpers/index.js';
+import contactsController from '../../controllers/contacts-controllers.js';
 
 const router = express.Router();
 
-const contactsAddSchema = Joi.object({
-  name: Joi.string()
-    .alphanum()
-    .min(3)
-    .max(30)
-    .required()
-    .messages({ 'any.required': `missing required name field` }),
-  email: Joi.string()
-    .required()
-    .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net'] } })
-    .messages({ 'any.required': `missing required email field` }),
-  phone: Joi.string()
-    .required()
-    .messages({ 'any.required': `missing required phone field` }),
-});
+router.get('/', contactsController.getContactsList);
 
-router.get('/', async (req, res, next) => {
-  try {
-    const result = await contactsServise.listContacts();
-    res.json(result);
-  } catch (error) {
-    console.log(error);
-    next(error);
-  }
-});
+router.get('/:contactId', contactsController.getContactById);
 
-router.get('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsServise.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post('/', contactsController.addNewContact);
 
-router.post('/', async (req, res, next) => {
-  try {
-    const { error } = contactsAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await contactsServise.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete('/:contactId', contactsController.deleteContactById);
 
-router.delete('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsServise.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json({
-      message: 'contact deleted',
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put('/:contactId', async (req, res, next) => {
-  try {
-    const { error } = contactsAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, 'missing fields');
-    }
-    const { contactId } = req.params;
-    const result = await contactsServise.updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put('/:contactId', contactsController.updateContactById);
 
 export default router;
